requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via fsspec
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
attrs==25.3.0
    # via aiohttp
certifi==2025.8.3
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via wandb
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
datasets==4.0.0
    # via -r requirements.in
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
filelock==3.19.1
    # via
    #   datasets
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.59.1
    # via matplotlib
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2025.3.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via wandb
hf-xet==1.1.8
    # via huggingface-hub
huggingface-hub==0.34.4
    # via
    #   datasets
    #   tokenizers
    #   transformers
hydra-core==1.3.2
    # via -r requirements.in
idna==3.10
    # via
    #   requests
    #   yarl
jinja2==3.1.6
    # via torch
joblib==1.5.1
    # via scikit-learn
kiwisolver==1.4.9
    # via matplotlib
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.5
    # via -r requirements.in
mpmath==1.3.0
    # via sympy
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
networkx==3.5
    # via torch
numpy==2.3.2
    # via
    #   -r requirements.in
    #   contourpy
    #   datasets
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   transformers
omegaconf==2.3.0
    # via hydra-core
packaging==25.0
    # via
    #   datasets
    #   huggingface-hub
    #   hydra-core
    #   matplotlib
    #   transformers
    #   wandb
pandas==2.3.1
    # via
    #   -r requirements.in
    #   datasets
pillow==11.3.0
    # via matplotlib
platformdirs==4.4.0
    # via wandb
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==6.32.0
    # via wandb
pyarrow==21.0.0
    # via datasets
pydantic==2.11.7
    # via wandb
pydantic-core==2.33.2
    # via pydantic
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   datasets
    #   huggingface-hub
    #   omegaconf
    #   transformers
    #   wandb
regex==2025.7.34
    # via transformers
requests==2.32.5
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
    #   wandb
safetensors==0.6.2
    # via transformers
scikit-learn==1.7.1
    # via -r requirements.in
scipy==1.16.1
    # via scikit-learn
sentry-sdk==2.35.2
    # via wandb
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sympy==1.14.0
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.4
    # via transformers
torch==2.8.0
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
transformers==4.55.2
    # via -r requirements.in
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
    #   typing-inspection
    #   wandb
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
wandb==0.21.3
    # via -r requirements.in
xxhash==3.5.0
    # via datasets
yarl==1.20.1
    # via aiohttp
